+++ src/components/FacetOptionFormatter/FacetOptionFormatter.test.js
@@ -1,6 +1,4 @@
-import {
-  render,
-} from '@testing-library/react';
+import { render } from '@testing-library/react';
 
 import FacetOptionFormatter from './FacetOptionFormatter';